feat: cache rule text with infractions
This change future-proofs infractions in the event that server rules are changed or deleted, causing infraction IDs that were once valid to now point to incorrect rules.

This commit introduces a fundamental breaking change to the database. A nullable TEXT column named RuleText *must* be added to the Infraction table of hammer.db